Main Refactor & New Tests
