+++ src/main/java/exceptions/CorrectInfoComparisonException.java
@@ -1,4 +1,4 @@
-package repositories;
+package exceptions;
 
 public class CorrectInfoComparisonException extends RuntimeException {
     public CorrectInfoComparisonException(String message) {
