+++ src/main/java/exceptions/WrongMenuChoiceException.java
@@ -1,4 +1,4 @@
-package ui;
+package exceptions;
 
 public class WrongMenuChoiceException extends RuntimeException {
     public WrongMenuChoiceException(String message) {
