+++ src/main/java/exceptions/WrongMenuFormatException.java
@@ -1,4 +1,4 @@
-package ui;
+package exceptions;
 
 public class WrongMenuFormatException extends Exception{
     public WrongMenuFormatException (String message){
